+++ Numerics/Geometry/Point3l.cs
@@ -801,7 +801,7 @@ namespace Ibasa.Numerics.Geometry
 		/// <param name="value1">The first point.</param>
 		/// <param name="value2">The second point.</param>
 		/// <returns>The distance between value1 and value2.</returns>
-		public static float Distance(Point3l value1, Point3l value2)
+		public static double Distance(Point3l value1, Point3l value2)
 		{
 			return Vector.Absolute(value2 - value1);
 		}
